requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
beautifulsoup4==4.8.2
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
django-url-filter==0.3.15
djangorestframework==3.11.0
enum-compat==0.0.3
Flask==1.1.1
httpie==2.0.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonfield==2.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
phonenumbers==8.11.2
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
